+++ changelog.md
@@ -6,6 +6,7 @@
 * Added new `proxyType` and `streamDelimiter` variables to Flash video to support RTMPS compatibility @rafa8626
 * Fixed issues with renderers when trying to use `muted` property while using `MediaElement` shim @rafa8626
 * Fixed typo with `Twitch` renderer related to trigger mouse events @rafa8626
+* Fixed typo when assigning options to `Dailymotion` renderer @rafa8626
 
 *4.2.6 (2017/09/19)*